Addressing user feedback on MIME multipart support programming model (see http://aspnetwebstack.codeplex.com/workitem/53).
Support for reading MIME multipart messages as handled by the HttpContent.ReadAsMultipartAsync extension methods had two usability problems in the original design:

1) The result was IEnumerable<HttpContent> which is too low-level to facilitate easy consumption. For example, it doesn't allow for converting multipart html form data into NameValueCollection data or for any other higher-level abstraction on top of the raw IEnumerable<HttpContent>

2) The stream provider doesn't enable any post processing of the data such as converting the read data into another format before the read task completed. This makes it impossible to provide a higher-level abstraction on top of IEnumerable<HttpContent> as the StreamProvider does not know when the message has been completely read.

This check-in fixes these short comings by first changing the signature of the ReadAsMultipartAsync from

    public static Task<IEnumerable<HttpContent>> ReadAsMultipartAsync(this HttpContent content, IMultipartStreamProvider streamProvider)

to

    public static Task<T> ReadAsMultipartAsync<T>(this HttpContent content, T streamProvider) where T : MultipartStreamProvider

It is now possible to write a MultipartStreamProvider that is completely tailored to the type of MIME multipart that it can read and present the result in the optimal way to the user.

Second, we introduce a “post processing” step as part of the MultipartStreamProvider that allows the implementation to do whatever post processing it wants on the MIME multipart body parts. For example, the MultipartFormDataStreamProvider implementation reads the FORM data parts and adds them to a NameValueCollection so they are easy to get at from the caller.

As part of this check-in we remove the helper functions for manipulate IEnumerable<HttpContent>. The functionality is now directly supported by the various MultipartStreamProvider implementations including MultipartFormDataStreamProvider, MultipartFileStreamProvider, and MultipartRelatedStreamProvider.