Performance improvements around simple parameters from the query string.  This gets about a 2x win in the case of calling get(int,int,int) in an in-memory loop.
Reorder the value provider factory and model binder providers to put the most common at the top.
Put TypeConverter and QueryStringProvider at the top. If we only pull from the query string (and have no misses), this avoids even invoking the other model binders or the route value provider.

Per-request caching QueryString and RouteValue providers so that we can reuse them across parameters. We shouldn't be parsing the query string for every parameter in a request.
Use array instead of IEnumerable. Avoids expensive ToList with ValueProviderFactories. No downside here and it gets about a 15% improvement in scenario.

Avoid ToList() calls in composite model binder. If we already have the array, use that.