Optional parameter fixes There're currently two issues with our optional parameter support 1. The default value on the optional parameter is not applied when using FormatterParameterBinding 2. We don't recognize an optional parameter when the default value is null (tracked by http://aspnetwebstack.codeplex.com/workitem/148)
For #1, we've decided not to support it by throwing. It's relatively uncommon to have optional parameter for the body and the formatters should be the ones deciding what to return for the default values.
For #2, we're introducing IsOptional concept to HttpParameterDescriptor so we're not relying on the default value because it can be null for optional parameters.