271: Progress Notifications should support more than 2G of bytes sent/received
In progress notifications, bytes received and bytes sent should be Int64 and not Int32 as they contain the cumulative values of the number of bytes received or sent and not just the number immediately received or sent. As it is fully possible to upload/download more than 2G of data with Web API this should be changed to Int64 to match how we deal with content-length in general (always as Int64).