Issue 254: HttpClient.PostAsJsonAsync and HttpClient.PostAsXmlAsync should set matching Accept header and have overload that takes System.Uri.
This commit adds missing overloads taking a System.Uri as HTTP request URI in addition to a string representation.

After further consideration we choose not to add accept header fields to match the request type so that PostAsJsonAsync sends an accept header of 'application/json' and PostAsXmlAsync sends an accept header of 'application/xml'. The reason is that some servers may react differently to an accept header, for example by returning a 406 status code instead. So rather than adding it implicitly, users can add it themselves if they so desire.