ODataConventionModelBuilder fixes.
1) added query composition mode for the convnetion model builder. this is
a special mode in which the model builder treats all types as entity types
so that we dont throw for collection properties. Also, this mode ignores
primitive collection types as they cannot be queried on anyways and we
dont support them.
2) Fixed an issue in which conventions are overriding user configuration
of required/optional property.
3) Improved error messages across model builder.
4) Fixed the KeyNotFound and KeyAlreadyExists issues when using generic
types in models. Type.FullName != Type.Namespace + Type.Name for generic
types which is triggering this issues. Added extensions methods EdmName
and EdmFullName on type which handle this issue.
5) Fixed an issue where we are re-adding ignored properties in convention
model builder.
6) Now that we have NotMappedAttribute, the model that we build might
contain types that are not used at all. So, added a new step
PruneUnReachableTypes() during model building.