Avoid loaded assemblies scan for edm types that we build.
When the ODataFormatter tries to find a clr type corresponding to an edm
type it first looks for ClrTypeAnnotation on the edm type. If it finds one
it will return the corresponding ClrType. Else it defaults to a full
assembly scan to find a clr type with the same full name as the edm type
which is intensive and should be avoided if possible.

This fix avoids it for edm types that the model builder builds by
pre-populating the annotation on the edm types that it builds.

Also caching models built in some querycomposition tests to improve test
runtimes.