Makes QueryableAttribute.ValidateQuery virtual to allow customization.
Addresses issue http://aspnetwebstack.codeplex.com/workitem/327

Prior to this change, ValidateQuery was a private static, and there was no way to override the default response behavior for invalid queries or to add support for new queries.

With this change, ValidateQuery is virtual and can be overridden to customize the response for invalid queries or to support new queries.

The existing behavior of ValidateQuery did not change.