Issue 350 and Issue 357.
1) Issue 357: Queryable attribute does not work with classes that has
properties in abstract parent classes.
Expression.Property doesn't work if the methodinfo that is passed is built
from the derived class and the property belongs to the base class.

2) Issue 350: QueryableAttribute can't filter on Get only property.
We are not building properties that have only non-public sets in to the
model. We should do that to support query composition on those properties.
But this would have the unfortunate side effect that the formatter would
fail when deserializing to this property which is fine.