Add ODataQuerySettings class to hold parameters to query composition.
Prior to this change, controlling null propagation or stable ordering was done through bool parameters to ApplyTo.  This increased the number of ApplyTo overloads and was not flexible for growth.

After this change, these bool parameters have been moved into ODataQuerySettings and the ApplyTo overloads reduced.  Future parameters to control query composition will be added to ODataQuerySettings as needed.