Fix for codeplex bug 710
Razor fails at runtime when encountering a comment containing a
single-quote inside a section block.

@section a { <!-- ' --> }

Fails at runtime with syntax error: Expected }

The root cause here is that section parsing has it's own method for
parsing a 'block' which doesn't correctly handle comments. Instead, the
parser treats the comment tag as a normal html tag. Since '!--' is not a
valid tag name it attempts to recover and find the end of the tag, it's
unable to find one so it just continues to the end of the document.

The safest fix is to add handling to the section parser for comments. I
considered merging the section parsing method and the general 'block'
parsing method, but this would introduce different semantics, and is not
desirable.

Adding support for xml processing instructions as well. This wasn't
reported, but should be there for completeness.

Note that it's still possible to encounter this syntax error when
providing incomplete markup inside a section block.

@section a { <a }

will result in the same syntax error. The 'recovery' process for tags
doesn't treat '}' as a recovery token, this would be a much larger change and
carries significant risk.