Issue 490879: multiple validation failures cause only one validation error in model state
This is a regression introduced while fixing TFS: 391851. The way it fixed
is wrong (https://aspnetwebstack.codeplex.com/SourceControl/changeset/9d359553f85333d201a083d83fef1b93d97de900)
as it is supressing multiple errors. It breaks the scenario where there
are multiple validation attributes present.

There is a codeplex issue opened for the same issue -
https://aspnetwebstack.codeplex.com/workitem/590 and this was fixed again in
a different and better manner.
https://aspnetwebstack.codeplex.com/SourceControl/changeset/353911090f0c02497d4003cd35a0effce43f0adf

Removing the earlier incorrect fix which caused this issue.